refactor: delete ContentViewConfigurationFactory extension
All logic has been moved into DocumentSurfaceViewModel.
File-drop methods moved into ContentView.
