+++ minimark/ContentView.swift
@@ -476,6 +476,66 @@ struct ContentView: View {
             surfaceViewModel.sourceMode = .web
         }
     }
+
+    private func handleDroppedFileURLs(_ fileURLs: [URL]) {
+        if let droppedFolderURL = ReaderFileRouting.firstDroppedDirectoryURL(from: fileURLs) {
+            guard folderWatchState.activeFolderWatch == nil else {
+                return
+            }
+
+            onAction(.requestFolderWatch(droppedFolderURL))
+            return
+        }
+
+        let markdownURLs = ReaderFileRouting.supportedMarkdownFiles(from: fileURLs)
+        guard !markdownURLs.isEmpty else {
+            return
+        }
+
+        let slotStrategy: FileOpenRequest.SlotStrategy =
+            document.fileURL == nil ? .reuseEmptySlotForFirst : .alwaysAppend
+        onAction(.requestFileOpen(FileOpenRequest(
+            fileURLs: markdownURLs,
+            origin: .manual,
+            slotStrategy: slotStrategy
+        )))
+    }
+
+    private func handlePickedFileURLs(_ fileURLs: [URL]) {
+        let markdownURLs = ReaderFileRouting.supportedMarkdownFiles(from: fileURLs)
+        guard !markdownURLs.isEmpty else {
+            return
+        }
+
+        let normalizedIncomingURL = ReaderFileRouting.normalizedFileURL(markdownURLs[0])
+        let currentURL = document.fileURL.map(ReaderFileRouting.normalizedFileURL)
+        if sourceEditing.hasUnsavedDraftChanges,
+           currentURL != normalizedIncomingURL {
+            onAction(.presentError(ReaderError.unsavedDraftRequiresResolution))
+            return
+        }
+
+        onAction(.requestFileOpen(FileOpenRequest(
+            fileURLs: [markdownURLs[0]],
+            origin: .manual,
+            slotStrategy: .replaceSelectedSlot
+        )))
+
+        let additionalMarkdownURLs = Array(markdownURLs.dropFirst())
+        guard !additionalMarkdownURLs.isEmpty else {
+            return
+        }
+
+        onAction(.requestFileOpen(FileOpenRequest(
+            fileURLs: additionalMarkdownURLs,
+            origin: .manual,
+            slotStrategy: .alwaysAppend
+        )))
+    }
+
+    private func canAcceptDroppedFileURLs(_ fileURLs: [URL]) -> Bool {
+        !ReaderFileRouting.containsLikelyDirectoryPath(in: fileURLs) || folderWatchState.activeFolderWatch == nil
+    }
 }
 
 struct DocumentSurfaceHost: View {
