--- minimark/Views/Content/ContentViewConfigurationFactory.swift
@@ -1,64 +1,0 @@
-import SwiftUI
-
-extension ContentView {
-
-    func handleDroppedFileURLs(_ fileURLs: [URL]) {
-        if let droppedFolderURL = ReaderFileRouting.firstDroppedDirectoryURL(from: fileURLs) {
-            guard folderWatchState.activeFolderWatch == nil else {
-                return
-            }
-
-            onAction(.requestFolderWatch(droppedFolderURL))
-            return
-        }
-
-        let markdownURLs = ReaderFileRouting.supportedMarkdownFiles(from: fileURLs)
-        guard !markdownURLs.isEmpty else {
-            return
-        }
-
-        let slotStrategy: FileOpenRequest.SlotStrategy =
-            document.fileURL == nil ? .reuseEmptySlotForFirst : .alwaysAppend
-        onAction(.requestFileOpen(FileOpenRequest(
-            fileURLs: markdownURLs,
-            origin: .manual,
-            slotStrategy: slotStrategy
-        )))
-    }
-
-    func handlePickedFileURLs(_ fileURLs: [URL]) {
-        let markdownURLs = ReaderFileRouting.supportedMarkdownFiles(from: fileURLs)
-        guard !markdownURLs.isEmpty else {
-            return
-        }
-
-        let normalizedIncomingURL = ReaderFileRouting.normalizedFileURL(markdownURLs[0])
-        let currentURL = document.fileURL.map(ReaderFileRouting.normalizedFileURL)
-        if sourceEditing.hasUnsavedDraftChanges,
-           currentURL != normalizedIncomingURL {
-            onAction(.presentError(ReaderError.unsavedDraftRequiresResolution))
-            return
-        }
-
-        onAction(.requestFileOpen(FileOpenRequest(
-            fileURLs: [markdownURLs[0]],
-            origin: .manual,
-            slotStrategy: .replaceSelectedSlot
-        )))
-
-        let additionalMarkdownURLs = Array(markdownURLs.dropFirst())
-        guard !additionalMarkdownURLs.isEmpty else {
-            return
-        }
-
-        onAction(.requestFileOpen(FileOpenRequest(
-            fileURLs: additionalMarkdownURLs,
-            origin: .manual,
-            slotStrategy: .alwaysAppend
-        )))
-    }
-
-    func canAcceptDroppedFileURLs(_ fileURLs: [URL]) -> Bool {
-        !ReaderFileRouting.containsLikelyDirectoryPath(in: fileURLs) || folderWatchState.activeFolderWatch == nil
-    }
-}
